requirements.txt
gunicorn
whitenoise
dj_database_url
django-bootstrap-v5
django-allauth
django-on-heroku
numpy
pandas
